+++ apps/deployer/test/status/application_test.exs
@@ -161,6 +161,9 @@ defmodule Deployer.Status.ApplicationTest do
     end)
     |> expect(:list, 1, fn -> [sname_1, sname_2, sname_3] end)
 
+    Foundation.RpcMock
+    |> stub(:call, fn _node, _module, _functions, _args, _timeout -> nil end)
+
     # No info, update needed
     assert {:noreply, %{monitoring: monitoring}} =
              StatusApp.handle_info(:update_apps, %{monitoring: []})
@@ -212,6 +215,9 @@ defmodule Deployer.Status.ApplicationTest do
 
     Deployer.HotUpgradeMock
     |> stub(:connect, fn _node -> {:error, :not_connecting} end)
+
+    Foundation.RpcMock
+    |> stub(:call, fn _node, _module, _functions, _args, _timeout -> nil end)
 
     # No info, update needed
     assert {:noreply, %{monitoring: monitoring}} =
@@ -268,6 +274,9 @@ defmodule Deployer.Status.ApplicationTest do
     Deployer.HotUpgradeMock
     |> stub(:connect, fn _node -> {:ok, :connected} end)
 
+    Foundation.RpcMock
+    |> stub(:call, fn _node, _module, _functions, _args, _timeout -> nil end)
+
     # No info, update needed
     assert {:noreply, %{monitoring: monitoring}} =
              StatusApp.handle_info(:update_apps, %{monitoring: []})
@@ -323,6 +332,9 @@ defmodule Deployer.Status.ApplicationTest do
 
     Deployer.HotUpgradeMock
     |> stub(:connect, fn _node -> {:ok, :connected} end)
+
+    Foundation.RpcMock
+    |> stub(:call, fn _node, _module, _functions, _args, _timeout -> nil end)
 
     ghosted_version = "1.1.1"
     version_map = StatusApp.current_version_map(sname_1)
@@ -402,6 +414,9 @@ defmodule Deployer.Status.ApplicationTest do
     end)
     |> stub(:list, fn -> [sname_1, sname_2, sname_3] end)
 
+    Foundation.RpcMock
+    |> stub(:call, fn _node, _module, _functions, _args, _timeout -> nil end)
+
     assert {:ok, _pid} = StatusApp.start_link(update_apps_interval: 50, name: module_name)
 
     {:ok, _map} = StatusApp.set_mode(module_name, name_1, :manual, "1.0.2")
@@ -452,6 +467,9 @@ defmodule Deployer.Status.ApplicationTest do
       }
     end)
     |> stub(:list, fn -> [sname_1, sname_2, sname_3] end)
+
+    Foundation.RpcMock
+    |> stub(:call, fn _node, _module, _functions, _args, _timeout -> nil end)
 
     assert {:ok, _pid} = StatusApp.start_link(update_apps_interval: 50, name: module_name)
 
@@ -504,6 +522,9 @@ defmodule Deployer.Status.ApplicationTest do
     end)
     |> stub(:list, fn -> [sname_1, sname_2, sname_3] end)
 
+    Foundation.RpcMock
+    |> stub(:call, fn _node, _module, _functions, _args, _timeout -> nil end)
+
     assert {:ok, _pid} = StatusApp.start_link(update_apps_interval: 50, name: module_name)
 
     {:ok, _map} = StatusApp.set_mode(module_name, name_1, :manual, "invalid")
@@ -546,6 +567,9 @@ defmodule Deployer.Status.ApplicationTest do
       }
     end)
     |> stub(:list, fn -> [sname_1, sname_2, sname_3] end)
+
+    Foundation.RpcMock
+    |> stub(:call, fn _node, _module, _functions, _args, _timeout -> nil end)
 
     assert {:ok, _pid} = StatusApp.start_link(update_apps_interval: 50, name: module_name)